WIP API refactor to make it slightly more consistent and predictable, bug fixes, updating/fixing example
