+++ Cargo.toml
@@ -1,11 +1,11 @@
 [workspace]
 members = [
-    #"crates/example",
+    "crates/example",
     #"crates/example-wasm",
     "crates/loading-bytes",
     "crates/renderling",
     #"crates/renderling-gpui",
-    "crates/sandbox"
+    #"crates/sandbox"
 ]
 
 exclude = ["./shaders"]
