+++ crates/renderling/Cargo.toml
@@ -108,7 +108,6 @@ gltf = {workspace = true, optional = true}
 half = "2.3"
 image = {workspace = true, features = ["hdr"]}
 log = {workspace = true}
-moongraph = { version = "0.4.2", default-features = false, features = ["dot", "derive"] }
 rustc-hash = "1.1"
 send_wrapper = "0.6"
 snafu = {workspace = true}
