+++ crates/renderling/src/stage/gltf_support.rs
@@ -6,14 +6,15 @@ use glam::{Mat4, Vec2, Vec3, Vec4};
 use snafu::{OptionExt, ResultExt, Snafu};
 
 use crate::{
+    atlas::{AtlasImage, AtlasTexture, RepackPreview, TextureAddressMode, TextureModes},
     pbr::{
         light::{DirectionalLight, Light, LightStyle, PointLight, SpotLight},
         Material,
     },
     slab::*,
     stage::Vertex,
-    AtlasImage, AtlasTexture, Camera, NestedTransform, Renderlet, RepackPreview, Stage,
-    TextureAddressMode, TextureModes, Transform,
+    stage::{NestedTransform, Renderlet, Stage},
+    Camera, Transform,
 };
 
 #[derive(Debug, Snafu)]
@@ -595,7 +596,7 @@ pub struct GltfNode {
 
 impl GltfNode {
     pub fn global_transform(&self) -> Transform {
-        todo!()
+        self.transform.get_global_transform()
     }
 }
 
@@ -948,6 +949,14 @@ impl Stage {
         GltfDocument::from_gltf(self, &document, buffers, images)
     }
 
+    pub fn load_gltf_document_from_bytes(
+        &mut self,
+        bytes: impl AsRef<[u8]>,
+    ) -> Result<GltfDocument, StageGltfError> {
+        let (document, buffers, images) = gltf::import_slice(bytes)?;
+        GltfDocument::from_gltf(self, &document, buffers, images)
+    }
+
     /// Draws the `StagedGltfNode` with the given `Camera`.
     pub fn draw_gltf_node(
         &mut self,
@@ -1008,8 +1017,8 @@ impl Stage {
 mod test {
     use crate::{
         pbr::{Material, PbrConfig},
-        stage::Vertex,
-        Camera, Context, Renderlet, Transform,
+        stage::{Renderlet, Vertex},
+        Camera, Context, Transform,
     };
     use crabslab::{GrowableSlab, Id, Slab};
     use glam::{Vec2, Vec3, Vec4, Vec4Swizzles};
