+++ crates/renderling/src/state.rs
@@ -20,12 +20,6 @@ pub enum WgpuStateError {
     CreateSurface { source: wgpu::CreateSurfaceError },
 }
 
-impl From<WgpuStateError> for moongraph::GraphError {
-    fn from(value: WgpuStateError) -> Self {
-        moongraph::GraphError::other(value)
-    }
-}
-
 pub enum RenderTarget {
     Surface {
         surface: wgpu::Surface<'static>,
